+++ src/pytest_randomly.py
@@ -1,9 +1,9 @@
 import argparse
+import hashlib
 import random
 import sys
-import time
 
-from pytest import Collector
+from pytest import Collector, fixture
 
 if sys.version_info >= (3, 8):
     from importlib.metadata import entry_points
@@ -46,7 +46,7 @@ except ImportError:
     have_numpy = False
 
 
-default_seed = int(time.time())
+default_seed = random.Random().getrandbits(32)
 
 
 def seed_type(string):
@@ -70,7 +70,7 @@ def pytest_addoption(parser):
         type=seed_type,
         help="""Set the seed that pytest-randomly uses (int), or pass the
                 special value 'last' to reuse the seed from the previous run.
-                Default behaviour: use int(time.time()), so the seed is
+                Default behaviour: use random.Random().getrandbits(32), so the seed is
                 different on each run.""",
     )
     group._addoption(
@@ -141,11 +141,12 @@ def _reseed(config, offset=0):
         faker_random.setstate(random_states[seed])
 
     if have_numpy:
-        if seed not in np_random_states:
-            np_random.seed(seed)
-            np_random_states[seed] = np_random.get_state()
+        numpy_seed = _truncate_seed_for_numpy(seed)
+        if numpy_seed not in np_random_states:
+            np_random.seed(numpy_seed)
+            np_random_states[numpy_seed] = np_random.get_state()
         else:
-            np_random.set_state(np_random_states[seed])
+            np_random.set_state(np_random_states[numpy_seed])
 
     if entrypoint_reseeds is None:
         entrypoint_reseeds = [
@@ -153,6 +154,15 @@ def _reseed(config, offset=0):
         ]
     for reseed in entrypoint_reseeds:
         reseed(seed)
+
+
+def _truncate_seed_for_numpy(seed):
+    seed = abs(seed)
+    if seed <= 2 ** 32 - 1:
+        return seed
+
+    seed_bytes = seed.to_bytes(seed.bit_length(), "big")
+    return int.from_bytes(hashlib.sha512(seed_bytes).digest()[: 32 // 8], "big")
 
 
 def pytest_report_header(config):
@@ -239,3 +249,10 @@ def reduce_list_of_lists(lists):
     for list_ in lists:
         new_list.extend(list_)
     return new_list
+
+
+if have_faker:
+
+    @fixture(autouse=True)
+    def faker_seed(pytestconfig):
+        return pytestconfig.getoption("randomly_seed")